requirements.txt
PyQt5>=5.15
pytesseract>=0.3
Pillow>=9.0
xlrd>=2.0
pyperclip>=1.8
